+++ homeworks/react-basics/rosyurch/public/manifest.json
