+++ homeworks/react-basics/rosyurch/src/Lecture.css
@@ -2,12 +2,6 @@ body {
     font-family: Arial, Helvetica, sans-serif;
 }
 
-.container {
-    /* display: flex;
-    flex: column wrap;
-    align-items: flex-start; */
-}
-
 .lecture {
     border: 1px solid #bbb;
 }
